:recycle: Make icon div creation a common function
